feat(SCC): moved `scc.h` and `slot.h` into libmsx.
