+++ src/slot_is_ROM.c
