+++ src/slot_is_ROM_p1.c
