+++ src/slot_is_ROM_p2.c
