+++ src/slot_is_SCC.c
