+++ src/slot_is_SCCPlus.c
