+++ src/slot_is_internal_OPLL.c
